move lang event to a new "asset" package
